 requirements.txt
# Xử lý PDF
PyMuPDF==1.22.5

# Embedding & ML
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.5

# FAISS
faiss-cpu==1.7.4

# Web server
Flask==2.3.3
requests==2.31.0

# NLP
nltk==3.8.1

# Google Gemini API
google-generativeai==0.1.2